+++ packages/strategies-keep3r/utils/v2-ftm-strategies.ts
@@ -70,6 +70,51 @@ export const v2FtmHarvestStrategies: v2FtmStrategy[] = [
     added: true,
     address: '0xbBdc83357287a29Aae30cCa520D4ed6C750a2a11',
   },
+  {
+    name: 'levscream_dai',
+    added: true,
+    address: '0xdC9D3bB76Df8bE1e9Ca68C7BF32293a86C829D81',
+  },
+  {
+    name: 'levscream_usdc',
+    added: true,
+    address: '0x42639b59cf9db5339C1C5cfB5738D0ba4F828F94',
+  },
+  {
+    name: 'levscream_mim',
+    added: true,
+    address: '0x8A807b5742587024783Df3Ed2F149725C197b5eE',
+  },
+  {
+    name: 'levscream_weth',
+    added: true,
+    address: '0x9DF9418837281faC4C4c1166f55D35F2799A9107',
+  },
+  {
+    name: 'levscream_wbtc',
+    added: true,
+    address: '0x0ed5C4effED69B39C86a8D806462d709Fb96A9E4',
+  },
+  {
+    name: 'levscream_crv',
+    added: true,
+    address: '0x6EEb47BBcDf0059E5F1D6Ee844Ba793D5401bF18',
+  },
+  {
+    name: 'levscream_spell',
+    added: true,
+    address: '0x35B51a621d78609dE7Cf25BC4e0682c7DEA38799',
+  },
+  {
+    name: 'levscream_dola',
+    added: true,
+    address: '0x36A1E9dF5EfdAB9694de5bFe25A9ACc23F66BCB7',
+  },
+  {
+    name: 'levscream_fusdt',
+    added: true,
+    address: '0x3B9bc1f596844Ca8a30A7710Aac7155968Db7d13',
+  },
 ];
 
 export const v2FtmTendStrategies: v2FtmStrategy[] = [];