+++ packages/stealth-txs/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/yearn/hardhat-monorepo/compare/@yearn/stealth-txs@0.1.0...@yearn/stealth-txs@0.2.0) (2022-01-12)
+
+
+### Bug Fixes
+
+* network utils and watcher aliveness ([fca2019](https://github.com/yearn/hardhat-monorepo/commit/fca2019d5ba68db207a3524a878518dc8b39f9c1))
+
+
+### Features
+
+* add eoa-gas toggle on stealth vault ([#25](https://github.com/yearn/hardhat-monorepo/issues/25)) ([31bd249](https://github.com/yearn/hardhat-monorepo/commit/31bd249ccd7b575235211dad447a856457e5427a))
+* reports hash via flashbots ([#21](https://github.com/yearn/hardhat-monorepo/issues/21)) ([de88bae](https://github.com/yearn/hardhat-monorepo/commit/de88bae1623e4749f5c15bc72500f23665920a55))
+* **stealth-txs:** adds pm2 to mainnet watcher script ([#53](https://github.com/yearn/hardhat-monorepo/issues/53)) ([7d0d09c](https://github.com/yearn/hardhat-monorepo/commit/7d0d09ccc0183a2e9e5ce146e9456924ea16e06a))
+
+
+
+
+
 # 0.1.0 (2021-11-08)