+++ packages/strategies-keep3r/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/yearn/hardhat-monorepo/compare/@yearn/strategies-keep3r@0.1.0...@yearn/strategies-keep3r@0.2.0) (2022-01-12)
+
+
+### Features
+
+* add 10 percent gaslimit buffer to detached work txs ([250a182](https://github.com/yearn/hardhat-monorepo/commit/250a182b2746173451b4a4dd2511fafb9edde5ad))
+* add 3 ftm starts to detached harvest ([013570d](https://github.com/yearn/hardhat-monorepo/commit/013570d2022ef0444a2aed4d1c2beffd627866eb))
+* add eoa-gas toggle on stealth vault ([#25](https://github.com/yearn/hardhat-monorepo/issues/25)) ([31bd249](https://github.com/yearn/hardhat-monorepo/commit/31bd249ccd7b575235211dad447a856457e5427a))
+* add ftm on strategies package ([6e31b3f](https://github.com/yearn/hardhat-monorepo/commit/6e31b3f83ce3fef89a76c6ac28273acef04d4e45))
+* add ftm strats ([#48](https://github.com/yearn/hardhat-monorepo/issues/48)) ([cc0f012](https://github.com/yearn/hardhat-monorepo/commit/cc0f0121a3fcb015d5d7c1bed736f9fbe005e8e8))
+* add levscream strats ([#61](https://github.com/yearn/hardhat-monorepo/issues/61)) ([20cbdee](https://github.com/yearn/hardhat-monorepo/commit/20cbdee904ac825501d341d30c5e2601582e611d))
+* add stealthRelayerCheck flag to safe guard ([#57](https://github.com/yearn/hardhat-monorepo/issues/57)) ([c1e355d](https://github.com/yearn/hardhat-monorepo/commit/c1e355d2efa71a2143ecc4a6c93a01cf7c9b02c8))
+* deployed keep3r StealthSafeGuard ([4de69b1](https://github.com/yearn/hardhat-monorepo/commit/4de69b128ab759c9251d981e483d4a0de941d916))
+* ftm to fantom ([#54](https://github.com/yearn/hardhat-monorepo/issues/54)) ([90bb5cf](https://github.com/yearn/hardhat-monorepo/commit/90bb5cf334e83efee512f3dd477dfe9badf958aa))
+* manual detached work script update ([#28](https://github.com/yearn/hardhat-monorepo/issues/28)) ([3c6fbc7](https://github.com/yearn/hardhat-monorepo/commit/3c6fbc720da98c0fd8e6ab8fdee2f37cd435bde2))
+* new ftm harvest levscreams strats ([#72](https://github.com/yearn/hardhat-monorepo/issues/72)) ([c6f6329](https://github.com/yearn/hardhat-monorepo/commit/c6f6329391c1913a84ad07be261974ae3bc11fcb))
+* remove deprecated stealth harvest strategies ([6c12657](https://github.com/yearn/hardhat-monorepo/commit/6c1265774ee84bddb48f7ebddfa06d1a7c3e8dde))
+* stealth safe guard add executors list ([d41851c](https://github.com/yearn/hardhat-monorepo/commit/d41851cbbb51eb5bc31f5a84fb3305fad252d5e0))
+
+
+
+
+
 # 0.1.0 (2021-11-08)